http: //ci2i-unstable.northscale.in/gsi-06.06.2024-23.47.pass.html

+++ secondary/vector/faiss/index_ivf.go
@@ -23,7 +23,7 @@ func (idx *IndexImpl) SetNProbe(nprobe int32) {
 	if ivfPtr == nil {
 		return
 	}
-	C.faiss_IndexIVF_set_nprobe(ivfPtr, C.ulong(nprobe))
+	C.faiss_IndexIVF_set_nprobe(ivfPtr, C.size_t(nprobe))
 }
 
 //Quantizer returns the pointer to the quantizer for
